major refactor
+ nuke a bunch of unused/old code
+ split all of the models into individual files
+ rejigger all of the imports
+ move extend, clone, complete methods onto the contract class
+ remove a bunch of integration tests and replace them with unit tests
+ fixes 370
