+++ purchasing/opportunities/forms.py
@@ -16,7 +16,7 @@ from purchasing.opportunities.models import Vendor
 
 from purchasing.utils import RequiredIf
 from purchasing.users.models import User, department_query
-from purchasing.data.models import opportunity_type_query
+from purchasing.data.contracts import ContractType
 
 ALL_INTEGERS = re.compile('[^\d.]')
 DOMAINS = re.compile('@[\w.]+')
@@ -162,7 +162,7 @@ class OpportunityForm(Form):
         allow_blank=True, blank_text='-----'
     )
     opportunity_type = QuerySelectField(
-        query_factory=opportunity_type_query,
+        query_factory=ContractType.opportunity_type_query,
         get_pk=lambda i: i.id,
         get_label=lambda i: i.name,
         allow_blank=True, blank_text='-----'
